chore: improve mocks
